+++ lib/models/dto/ui_config_dto.dart
@@ -3,16 +3,19 @@ import 'dart:convert';
 class UiConfigDTO {
   UiConfigDTO({
     this.itemTabShowDebt = true,
-    this.itemTabShowDailyStats = true,
+    this.itemTabShowDailyBar = true,
+    this.itemTabShowDailyCalendar = true,
   });
 
   final bool itemTabShowDebt;
-  final bool itemTabShowDailyStats;
+  final bool itemTabShowDailyBar;
+  final bool itemTabShowDailyCalendar;
 
   static UiConfigDTO _fromJson(Map<String, dynamic> json) {
     return UiConfigDTO(
       itemTabShowDebt: json['itemTabShowDebt'] ?? true,
-      itemTabShowDailyStats: json['itemTabShowDailyStats'] ?? true,
+      itemTabShowDailyBar: json['itemTabShowDailyBar'] ?? true,
+      itemTabShowDailyCalendar: json['itemTabShowDailyCalendar'] ?? true,
     );
   }
 
@@ -23,7 +26,8 @@ class UiConfigDTO {
   static Map<String, dynamic> _toJson(UiConfigDTO uiConfig) {
     return {
       'itemTabShowDebt': uiConfig.itemTabShowDebt,
-      'itemTabShowDailyStats': uiConfig.itemTabShowDailyStats,
+      'itemTabShowDailyBar': uiConfig.itemTabShowDailyBar,
+      'itemTabShowDailyCalendar': uiConfig.itemTabShowDailyCalendar,
     };
   }